+++ vitest.config.ts
@@ -1,8 +1,8 @@
-import { defineConfig } from "vitest/config";
+import { defineConfig } from 'vitest/config'
 
 export default defineConfig({
-  test: {
-    outputFile: "tmp/test-results.json",
-    include: ["src/**/*.test.ts"],
-  },
-});
+	test: {
+		outputFile: 'tmp/test-results.json',
+		include: ['src/**/*.test.ts']
+	}
+})